releases: fix versions tests + adds folders structure
