+++ tools/release/lib/tests/samples/ada_version_h_expected.txt
